Commit main and du (mid-stable)
The `du` function recursively calculates the total disk usage of
a given file ROOT. It performs a DFS through the file tree starting
from ROOT and going further down until no more directories are
present.

By default, disk usage is computed in kylobytes, hence dividing the
blocks allocated for a file by two.

There are still some issues that need to be addressed for this to
be considered a stable version:

1. Error reporting: Returning (uint64_t)-1 for errors conflates
   with possible valid return values since uint64_t is an unsigned
   type, making -1 a very large positive number, which could
   theoretically be a valid disk usage value.
2. Error Message Inconsitencies: The format and detail of error
   messages are inconsistent.
3. Command-Line Argument Parsing: Argument parsing logic allows for
   scenarios not intended by the requirements (e.g. passing two
   filenames).
4. Resource Management: The function du has multiple points where
   it returns early due to an error, each time duplicating the
   `closedir(dirp)` call.
5. Program behavior is incorrect when passing file as input.