Implement dynamic array to track seen I-nodes
Keeping track of seen i-nodes allows us to avoid double counting
hardlinks. The choice of data structure was made based on simplicity at
the expense of performance. Future improvements should be focused on
implementing a data structure with a more efficient lookup (perhaps a
set or an AVL tree).

Additional Notes:
- For consistency, PATH_MAX was renamed to kPathMax and, hence, limits.h
  is no longer included.
- Using uint64_t to represent type blkcnt_t seems unnecessary and
  counter-intuitive. All references to this type were changes
  accordingly and stint.h is no longer included.